tg_accounting_bot/requirements.txt
python-telegram-bot==20.0
SQLAlchemy
